streamlit_app/requirements.txt
streamlit
python-dotenv
groq
pypdf